move job assignment calls to job assignments controller
